Separate keywords table from the Lexer
Lexer.is_keyword was exposed as part of the support for raw identifier
syntax in OCaml 5.2. Using this function directly added a dependency
for Path and Pprintast on Lexer and transitively on Parser which did not
exist before. Unfortunately, this significantly reduces the parallelism
when compiling ocamlcommon.cmxa.

The fix here is to move the keyword_table from Lexer to a new module
Keywords and instead expose Keywords.is_keyword and additionally
Keywords.token_of_string for the Lexer itself.

There is a small wart that Parser.token is required, but can't be
directly referenced because this would create a dependency for the new
keywords.cmx on parser.cmx and undermine the fix. The solution is to
generate an additional mli-only module Keywords_token containing a copy
(with type equality) of Parser.token and use that.