WIP: compile type-checker in parallel
Introduce an extra module Typecore_ind for Typemod and Typeclass to use,
which allows them to be compiled at the same time as Typecore.

Trivially, in passing, Translcore can use Typecore_ind as well, which
allows most of the bytecode compiler to be compiled in parallel with the
type checker.